Switch from greedy meshing to reimplemented visible faces meshing

+++ src/chunk/position.rs
@@ -6,21 +6,21 @@ use super::ChunkData;
 
 #[derive(Component, Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord)]
 pub struct ChunkPos {
-    pub x: u32,
-    pub y: u32,
-    pub z: u32,
+    pub x: usize,
+    pub y: usize,
+    pub z: usize,
 }
 
 impl ChunkPos {
-    pub fn new(x: u32, y: u32, z: u32) -> Self {
+    pub fn new(x: usize, y: usize, z: usize) -> Self {
         Self { x, y, z }
     }
 
     pub fn from_global_coords(x: f32, y: f32, z: f32) -> Self {
         Self {
-            x: (x / (Voxel::size() * ChunkData::edge() as f32)) as u32,
-            y: (y / (Voxel::size() * ChunkData::edge() as f32)) as u32,
-            z: (z / (Voxel::size() * ChunkData::edge() as f32)) as u32,
+            x: (x / (Voxel::size() * ChunkData::edge() as f32)) as usize,
+            y: (y / (Voxel::size() * ChunkData::edge() as f32)) as usize,
+            z: (z / (Voxel::size() * ChunkData::edge() as f32)) as usize,
         }
     }
